TBP_Projekt_Luka_Hrupec/app/requirements.txt
flask
psycopg2-binary

